extract error classes
