+++ frontend/src/restful/HttpResponseError.js
@@ -1,0 +1,8 @@
+class HttpResponseError extends Error {
+  constructor(status) {
+    super(`got ${status}`);
+    this.status = status;
+  }
+}
+
+export default HttpResponseError;
